vminfo: Track VMs by UUID rather than ephemeral ID
Current VMWARE-VMINFO-MIB states for vmwVmVMID:
"No longer provided, use vmwVmIdx. See vmwVmUUID for cross system,
unique, persistent identifier."
and about vmwVmIdx:
"An operational identifier given the VM when registered on this ESX
system. The value is not unique across ESX systems and may
change upon reboot."

This means rebooting the ESX system may cause us to delete and re-create
all of the VMs!

The vmwVmUUID field was added 2010-06-22, and vmwVmVMID field
removed 2011-09-17 according to the revision history.

On libvirt, the story is similar but worse. The integer "Id" is only
present for running VMs, and it is also ephemeral which means we don't
actually track VMs correctly at all, even without a host reboot.
A real example:

$ virsh -c 'qemu:///system' list --all
 Id    Name                           State
----------------------------------------------------
 --snip--
 10    buildbot-linux                 running
 16    test                           running
 -     alarm-aarch64                  shut off

$ virsh -c 'qemu:///system' shutdown test
Domain test is being shutdown

$ virsh -c 'qemu:///system' list --all
 Id    Name                           State
----------------------------------------------------
--snip--
 10    buildbot-linux                 running
 -     alarm-aarch64                  shut off
 -     test                           shut off

$ virsh -c 'qemu:///system' start test
Domain test started

$ virsh -c 'qemu:///system' list --all
 Id    Name                           State
----------------------------------------------------
--snip--
 10    buildbot-linux                 running
 17    test                           running
 -     alarm-aarch64                  shut off

Note that the Id for test changed when it was stopped, then started.